requirements.txt
asgiref==3.2.10
Django==2.2.8
django-debug-toolbar==2.2
django-jsonfield==1.4.0
djangorestframework==3.11.1
hashids==1.3.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
